+++ code/app/Providers/RouteServiceProvider.php
@@ -18,6 +18,7 @@ use App\Models\Vote\Ballot;
 use App\Models\Vote\BallotCompletion;
 use App\Models\Wiki\Article;
 use App\Models\Wiki\ArticleIteration;
+use App\Providers\AtheniaProviders\BaseRouteServiceProvider;
 use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Route;
 
@@ -25,53 +26,16 @@ use Illuminate\Support\Facades\Route;
  * Class RouteServiceProvider
  * @package App\Providers
  */
-class RouteServiceProvider extends ServiceProvider
+class RouteServiceProvider extends BaseRouteServiceProvider
 {
     /**
-     * @var array
-     */
-    private $modelPlaceHolders = [
-        'article' => Article::class,
-        'article_iteration' => ArticleIteration::class,
-        'ballot' => Ballot::class,
-        'ballot_completion' => BallotCompletion::class,
-        'category' => Category::class,
-        'collection' => Collection::class,
-        'collection_item' => CollectionItem::class,
-        'feature' => Feature::class,
-        'membership_plan' => MembershipPlan::class,
-        'organization' => Organization::class,
-        'organization_manager' => OrganizationManager::class,
-        'payment_method' => PaymentMethod::class,
-        'role' => Role::class,
-        'subscription' => Subscription::class,
-        'user' => User::class,
-    ];
-
-    /**
-     * Define your route model bindings, pattern filters, etc.
+     * Gets all application specific model placeholders
      *
-     * @return void
+     * @return array
      */
-    public function boot()
+    public function getAppModelPlaceholders(): array
     {
-        foreach($this->modelPlaceHolders as $placeHolder => $model) {
-            Route::pattern($placeHolder, '^[0-9]+$');
-            Route::model($placeHolder, $model);
-        }
-
-        parent::boot();
-    }
-
-    /**
-     * Define the routes for the application.
-     *
-     * @return void
-     */
-    public function map()
-    {
-        Route::middleware('api-v1')
-            ->namespace('App\Http\V1\Controllers')
-            ->group(base_path('routes/api-v1.php'));
+        return [
+        ];
     }
 }
