+++ code/tests/Unit/Providers/AuthServiceProviderTest.php
@@ -6,12 +6,12 @@ namespace Tests\Unit\Providers;
 use App\Contracts\Repositories\User\UserRepositoryContract;
 use App\Models\User\User;
 use App\Policies\User\UserPolicy;
+use App\Providers\AtheniaProviders\AuthServiceProvider;
+use App\Services\UserAuthenticationService;
 use Illuminate\Auth\AuthManager;
 use Illuminate\Contracts\Auth\Access\Gate;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Hashing\Hasher;
-use App\Providers\AuthServiceProvider;
-use App\Services\UserAuthenticationService;
 use ReflectionClass;
 use Tests\TestCase;
 use Tests\Traits\ReflectionHelpers;
